Fix all compiler warnings and other changes
- Remove clone_expr function
- Use size_t everywhere appropriate instead of int
- Use longs for Lisp numbers, not ints

+++ src/intern.h
@@ -3,6 +3,8 @@
 #ifndef INTERN_H
 #define INTERN_H
 
+#include <stddef.h>
+
 typedef unsigned int InternID;
 
 // Interns the string and returns its unique identifier.
@@ -10,7 +12,7 @@ InternID intern_string(const char *str);
 
 // Interns a string of n characters. Does not require a null terminator. If the
 // same string has previously been interned, returns the same identifier.
-InternID intern_string_n(const char *str, int n);
+InternID intern_string_n(const char *str, size_t n);
 
 // Looks up a string by identifier.
 const char *find_string(InternID id);
